+++ src/model/neuralNetwork.hpp
@@ -9,7 +9,7 @@ namespace nn {
 struct neural_network {
 	std::vector<std::unique_ptr<Layer>> layers;
 	const NetworkConfig &config;
-	int getLayerCount() const { return (config.hidden_layer_count() + 1); }
+	size_t getLayerCount() const { return (config.hidden_layer_count() + 1); }
 	neural_network(const NetworkConfig &network_config);
 	void reset();
 	~neural_network() = default;
